EXTRA: Started the refactor pattern for moving pages out of the routes folder
