+++ services/ui-src/src/components/index.ts
@@ -21,6 +21,10 @@ export { PreviewBanner } from "./banners/PreviewBanner";
 export { Card } from "./cards/Card";
 export { EmailCard } from "./cards/EmailCard";
 export { TemplateCard } from "./cards/TemplateCard";
+// dashboard 
+export { DashboardPage } from "./pages/Dashboard/DashboardPage"
+export { DashboardList } from "./pages/Dashboard/DashboardProgramList";
+export { MobileDashboardList } from "./pages/Dashboard/DashboardProgramListMobile";
 // drawers
 export { Drawer } from "./drawers/Drawer";
 export { ReportDrawer } from "./drawers/ReportDrawer";
