+++ services/ui-src/src/routes/index.tsx
@@ -7,7 +7,4 @@ export { NotFound } from "./NotFound/NotFound";
 export { Profile } from "./Profile/Profile";
 // report
 export { GetStarted } from "./GetStarted/GetStarted";
-export { DashboardPage } from "./Dashboard/DashboardPage";
-export { DashboardList } from "./Dashboard/DashboardList";
-export { MobileDashboardList } from "./Dashboard/MobileDashboardList";
 export { ReviewSubmit } from "./ReviewSubmit/ReviewSubmit";
